Revert "Update grpc client (#29)"
This reverts commit 5c072885cd77aacaaf0f01b092200d5695f38dcc.
