+++ .gitignore
@@ -145,11 +145,3 @@ build/
 
 # version file is auto-generated
 src/version.ts
-
-### vscode ###
-.vscode
-!.vscode/settings.json
-!.vscode/tasks.json
-!.vscode/launch.json
-!.vscode/extensions.json
-*.code-workspace
