+++ package.json
@@ -53,8 +53,7 @@
     "eslint-plugin-prettier": "^4.2.1",
     "husky": "^8.0.1",
     "prettier": "^2.4.0",
-    "pretty-quick": "^3.1.3",
-    "readline-sync": "^1.4.10"
+    "pretty-quick": "^3.1.3"
   },
   "dependencies": {
     "@grpc/grpc-js": "^1.8.14",
