+++ scripts/build.sh
@@ -41,5 +41,4 @@ cp package.json build/
 cp README.md build/
 
 # Copy Proto Files
-cp -R ./src/proto ./build
-cp -R ./submodules/durabletask-protobuf/protos/orchestrator_service.proto ./build/proto
+cp -R ./submodules/durabletask-protobuf/protos ./build
