+++ src/index.ts
@@ -1,4 +1,4 @@
-import { TaskHubGrpcClient } from "./client/client";
+import { TaskHubGrpcClient } from "./client";
 import { TaskHubGrpcWorker } from "./worker/task-hub-grpc-worker";
 import { OrchestrationContext } from "./task/context/orchestration-context";
 import { ActivityContext } from "./task/context/activity-context";
