+++ src/task/context/orchestration-context.ts
@@ -43,7 +43,7 @@ export abstract class OrchestrationContext {
   abstract createTimer(fireAt: Date | number): Task<any>;
 
   /**
-   * Schedule an activity for execution.
+   * chedule an activity for execution.
    *
    * @param {Orchestrator} orchestrator The sub-orchestrator function to call.
    * @param {TInput} input The JSON-serializable input value for the sub-orchestrator function.
@@ -51,7 +51,7 @@ export abstract class OrchestrationContext {
    *
    * @returns {Task<TOutput>} A Durable Task that completes when the sub-orchestrator function completes.
    */
-  abstract callActivity<TInput, TOutput>(activity: TActivity<TInput, TOutput> | string, input?: TInput): Task<TOutput>;
+  abstract callActivity<TInput, TOutput>(activity: TActivity<TInput, TOutput>, input?: TInput): Task<TOutput>;
 
   /**
    * Schedule sub-orchestrator function for execution.
@@ -63,7 +63,7 @@ export abstract class OrchestrationContext {
    * @returns {Task<TOutput>} A Durable Task that completes when the sub-orchestrator function completes.
    */
   abstract callSubOrchestrator<TInput, TOutput>(
-    orchestrator: TOrchestrator | string,
+    orchestrator: TOrchestrator,
     input?: TInput,
     instanceId?: string,
   ): Task<TOutput>;
