+++ src/worker/runtime-orchestration-context.ts
@@ -241,10 +241,7 @@ export class RuntimeOrchestrationContext extends OrchestrationContext {
     return timerTask;
   }
 
-  callActivity<TInput, TOutput>(
-    activity: TActivity<TInput, TOutput> | string,
-    input?: TInput | undefined,
-  ): Task<TOutput> {
+  callActivity<TInput, TOutput>(activity: TActivity<TInput, TOutput>, input?: TInput | undefined): Task<TOutput> {
     const id = this.nextSequenceNumber();
     const name = typeof activity === "string" ? activity : getName(activity);
     const encodedInput = input ? JSON.stringify(input) : undefined;
@@ -257,17 +254,12 @@ export class RuntimeOrchestrationContext extends OrchestrationContext {
   }
 
   callSubOrchestrator<TInput, TOutput>(
-    orchestrator: TOrchestrator | string,
+    orchestrator: TOrchestrator,
     input?: TInput | undefined,
     instanceId?: string | undefined,
   ): Task<TOutput> {
-    let name;
-    if (typeof orchestrator === "string") {
-      name = orchestrator;
-    } else {
-      name = getName(orchestrator);
-    }
     const id = this.nextSequenceNumber();
+    const name = getName(orchestrator);
 
     // Create a deterministic instance ID based on the parent instance ID
     // use the instanceId and apprent the id to it in hexadecimal with 4 digits (e.g. 0001)
